# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface UploadUrlRequest {
  /**
     * @minLength 1
     * @maxLength 255
     */
  name: string;
  /**
     * @minimum 1
     * @maximum 262144000
     */
  size: number;
  /**
     * @minLength 1
     * @maxLength 100
     */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export type PostMediaType = typeof PostMediaType[keyof typeof PostMediaType];


export const PostMediaType = {
  image: 'image',
  video: 'video',
} as const;

export interface Post {
  id: string;
  caption: string;
  objectPath: string;
  mediaType: PostMediaType;
  fileName: string;
  authorId: string;
  authorName: string;
  /** @nullable */
  authorImage: string | null;
  createdAt: string;
}

export type PostInputMediaType = typeof PostInputMediaType[keyof typeof PostInputMediaType];


export const PostInputMediaType = {
  image: 'image',
  video: 'video',
} as const;

export interface PostInput {
  /** @maxLength 2200 */
  caption: string;
  /**
     * @minLength 1
     * @maxLength 512
     */
  objectPath: string;
  mediaType: PostInputMediaType;
  /**
     * @minLength 1
     * @maxLength 255
     */
  fileName: string;
}

export interface ErrorEnvelope {
  error: string;
}

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type LogoutBrowserSessionParams = {
returnTo?: string;
};