# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Browser authentication endpoints.
  - name: Posts
    description: Public photo and video posts.
  - name: Storage
    description: Media upload and serving endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the signed-in user, if any
      responses:
        "200":
          description: Current authentication state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start browser sign-in
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect to the sign-in provider.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete browser sign-in
      responses:
        "302":
          description: Redirect after sign-in.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: End the browser session
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect after sign-out.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MobileTokenExchangeRequest"
      responses:
        "200":
          description: Session created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MobileTokenExchangeSuccess"
        "400":
          description: Invalid authorization code exchange.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: No usable identity claims.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: End a mobile session
      responses:
        "200":
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutSuccess"
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a secure direct-upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Invalid upload metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Sign-in required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve a stored media file according to its access policy
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Media bytes.
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Object not found.
  /posts:
    get:
      tags: [Posts]
      operationId: listPosts
      summary: List public posts
      responses:
        "200":
          description: Public posts, newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      tags: [Posts]
      operationId: createPost
      summary: Publish an uploaded image or video
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Post created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          description: Invalid post data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Sign-in required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /posts/mine:
    get:
      tags: [Posts]
      operationId: listMyPosts
      summary: List posts owned by the signed-in user
      responses:
        "200":
          description: The signed-in user's posts, newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
        "401":
          description: Sign-in required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /posts/{id}:
    get:
      tags: [Posts]
      operationId: getPost
      summary: Get a public post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Public post.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Post not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      tags: [Posts]
      operationId: deletePost
      summary: Delete a post owned by the signed-in user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Post deleted.
        "401":
          description: Sign-in required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Post not found or not owned by the signed-in user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, "null"]
          format: email
        firstName:
          type: [string, "null"]
        lastName:
          type: [string, "null"]
        profileImageUrl:
          type: [string, "null"]
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        size:
          type: integer
          minimum: 1
          maximum: 262144000
        contentType:
          type: string
          minLength: 1
          maxLength: 100
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"
    Post:
      type: object
      required: [id, caption, objectPath, mediaType, fileName, authorId, authorName, authorImage, createdAt]
      properties:
        id:
          type: string
          format: uuid
        caption:
          type: string
        objectPath:
          type: string
        mediaType:
          type: string
          enum: [image, video]
        fileName:
          type: string
        authorId:
          type: string
        authorName:
          type: string
        authorImage:
          type: [string, "null"]
          format: uri
        createdAt:
          type: string
          format: date-time
    PostInput:
      type: object
      required: [caption, objectPath, mediaType, fileName]
      properties:
        caption:
          type: string
          maxLength: 2200
        objectPath:
          type: string
          minLength: 1
          maxLength: 512
        mediaType:
          type: string
          enum: [image, video]
        fileName:
          type: string
          minLength: 1
          maxLength: 255
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string