# lib/api-zod/src/generated/types/post.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PostMediaType } from './postMediaType';

export interface Post {
  id: string;
  caption: string;
  objectPath: string;
  mediaType: PostMediaType;
  fileName: string;
  authorId: string;
  authorName: string;
  /** @nullable */
  authorImage: string | null;
  createdAt: Date;
}