# lib/api-zod/src/generated/types/postInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PostInputMediaType } from './postInputMediaType';

export interface PostInput {
  /** @maxLength 2200 */
  caption: string;
  /**
     * @minLength 1
     * @maxLength 512
     */
  objectPath: string;
  mediaType: PostInputMediaType;
  /**
     * @minLength 1
     * @maxLength 255
     */
  fileName: string;
}