# lib/api-zod/src/generated/types/uploadUrlRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UploadUrlRequest {
  /**
     * @minLength 1
     * @maxLength 255
     */
  name: string;
  /**
     * @minimum 1
     * @maximum 262144000
     */
  size: number;
  /**
     * @minLength 1
     * @maxLength 100
     */
  contentType: string;
}